DalManager
Updated DalManager position and added CheckServer(), GetAllServers()
